+++ kompendium-core/src/main/kotlin/io/bkbn/kompendium/core/handler/MapHandler.kt
@@ -1,4 +1,4 @@
-package io.bkbn.kompendium.core.schema
+package io.bkbn.kompendium.core.handler
 
 import io.bkbn.kompendium.core.Kontent.generateKTypeKontent
 import io.bkbn.kompendium.core.Kontent.generateKontent
