+++ kompendium-core/src/main/kotlin/io/bkbn/kompendium/core/handler/SchemaHandler.kt
@@ -1,4 +1,4 @@
-package io.bkbn.kompendium.core.schema
+package io.bkbn.kompendium.core.handler
 
 import io.bkbn.kompendium.core.metadata.SchemaMap
 import kotlin.reflect.KClass
